+++ react/EmptyState.js
@@ -1,0 +1,3 @@
+import EmptyState from './components/EmptyState/index'
+
+export default EmptyState