+++ react/IconArrowBack.js
@@ -1,0 +1,3 @@
+import ArrowBack from './components/icon/ArrowBack/index'
+
+export default ArrowBack