+++ react/IconCaretDown.js
@@ -1,0 +1,3 @@
+import CaretDown from './components/icon/CaretDown/index'
+
+export default CaretDown