+++ react/IconCaretRight.js
@@ -1,3 +1,3 @@
-import CaretRight from './components/icon/CaretRight'
+import CaretRight from './components/icon/CaretRight/index'
 
 export default CaretRight