+++ react/IconCaretUp.js
@@ -1,0 +1,3 @@
+import CaretUp from './components/icon/CaretUp/index'
+
+export default CaretUp