+++ react/IconClock.js
@@ -1,0 +1,3 @@
+import Clock from './components/icon/Clock/index'
+
+export default Clock