+++ react/IconCog.js
@@ -1,0 +1,3 @@
+import Cog from './components/icon/Cog/index'
+
+export default Cog