+++ react/IconDelete.js
@@ -1,0 +1,3 @@
+import Delete from './components/icon/Delete/index'
+
+export default Delete