+++ react/IconExternalLink.js
@@ -1,0 +1,3 @@
+import ExternalLink from './components/icon/ExternalLink/index'
+
+export default ExternalLink