+++ react/IconExternalLinkMini.js
@@ -1,0 +1,3 @@
+import ExternalLinkMini from './components/icon/ExternalLinkMini/index'
+
+export default ExternalLinkMini