+++ react/IconHelp.js
@@ -1,0 +1,3 @@
+import Help from './components/icon/Help/index'
+
+export default Help